apps/workers/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-venture-architect-workers"
version = "0.1.0"
description = "Background workers for AI Venture Architect platform"
authors = [
    {name = "AI Venture Architect Team"}
]
dependencies = [
    "crewai>=0.22.0",
    "langgraph>=0.0.40",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "nats-py>=2.6.0",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "httpx>=0.25.2",
    "beautifulsoup4>=4.12.2",
    "feedparser>=6.0.10",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.2",
    "hdbscan>=0.8.33",
    "umap-learn>=0.5.4",
    "openai>=1.6.0",
    "anthropic>=0.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "ulid-py>=1.1.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/workers"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "I", "N", "UP", "ANN", "S", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["ANN101", "ANN102", "COM812", "ISC001"]
